==> step-01.01-descriptive-statistics.RMD <==
---
title: "Descriptive Stats"
author: "Jesse Lecy"
output:
  html_document:
    theme: readable
    df_print: paged
    highlight: zenburn
    toc: true
    self_contained: false
    number_sections: false
    css: clean.css
---







# SETUP

```{r, warning=FALSE, message=FALSE}

library( foreign )
library( memisc )
library( knitr )
library( dplyr )
library( xtable )
library( stargazer )
library( broom)
library( pander )



```



# LOAD DATA

```{r, warning=FALSE }


dat <- readRDS( "./Data/CompleteHazardSpells.rds" )

# lapply( dat, class )
                  
# head( dat, 25 ) %>% pander
```




# DESCRIPTIVE STATISTICS


## Table of Spells

```{r}

# num of times each org occurs in dataset

table( table( dat$ein)) %>% pander()

```

## Organizations by Year

```{r} 

table( dat$fisyr )  %>% pander

```

## when do orgs professionalize


```{r}

table( dat$fisyr, dat$prof ) %>% pander

round( prop.table( table( dat$fisyr, dat$prof ), margin=1 ), 3 ) %>% pander

```






# STARGAZER DESCRIPTIVES

```{r, results='asis'}

# results='asis'

# "max" maximum
# "mean" mean
# "median" median
# "min" minimum
# "n" number of observations
# "p25" 25th percentile
# "p75" 75th percentile
# "sd" standard deviation

dd <- dat[ , c("prof","Accrual","GovtMoneyRat","UNAgrand","FixedCostRat",
               "SurplusRat_ndrop_w892","EqRat_w_K","ProfFundFeeYes",
               "FS_Totrev_adj","HHI") ]


stargazer( dd, digits=4,
           type = "html", out="./Results/DescriptiveStatistics.doc", 
           title="Descriptive Statistics",
           summary.stat = c( "min", "median", "mean", "max", "sd" ) )



```
