==> step-03-calculate-effects.Rmd <==
---
title: "Step 03 - Create Tables"
author: "Jesse Lecy"
output:
  html_document:
    theme: readable
    df_print: paged
    highlight: zenburn
    toc: true
    self_contained: false
    number_sections: false
    css: clean.css
---



```{r, echo=F, warning=FALSE, message=FALSE}

library( foreign )
library( memisc )
library( knitr )
library( dplyr )
library( xtable)
library( stargazer)
library( broom)
library( pander)
library( plotly )
# library( ReporteRs )  # no longer available
library( officer )  # replacement for ReporteRs


```


# MATH OF EFFECT SIZE

Here we define effect size as the increase or decrease in probability of professionalization when a specific covariate is present.

Typically we look at the median time to success/failure (professionalization) as the average effect, and then compare all effects to the average. In this case, we don't have a long enough panel to calculate the median time to success, so we have a choice:

* Use the model that fits
* Compare cases at age=5 since this is the last data point we have.

$$ h_j = \frac{1}{1 + e^{-(a_1D_1 + a_2D_2 + ... + a_jD_j + X\beta ) }}  $$


The baseline hazard rate then for any particular time j is calculated as:

$$ h_j = \frac{1}{1 + e^{-(a_jD_j ) }} $$

And the effect size of a binary covariate at that period would be:

$$ h_j = \frac{1}{1 + e^{-(a_jD_j + \beta_1 \cdot ProfessionalFundraising) }} $$

Where you compare the case where ProfFundraising=1 to the case where ProfFundraising=0.

For continuous covariates, like HHI, you would compare the case where the covariate equals the mean, to a case where the covariate is at the mean plus a standard deviation, or conversely a case at the 25th percentile (low HHI) to 75th percentile (high HHI).

The effect in these cases will represent the change in probability of professionalization as a result of the covariate.



# LOAD DATA

```{r, warning=FALSE }


dat <- readRDS( "./Data/CompleteHazardSpells.rds" )

# copy original dataset

dat.original <- dat

```



# BASELINE CASE



## Baseline Model

```{r}


m.all <- glm( formula= prof ~ D1 + D2 + D3 + D4 + D5 + D6 + Accrual + GovtMoneyRat + HHI + 
                             UNAgrand + FixedCostRat + SurplusRat_ndrop_w892 + EqRat_w_K + 
                             ProfFundFeeYes - 1, 
                             data=dat,   family=binomial("logit") )

m.all %>% tidy %>% kable( digits=3 )


```


## Create Baseline Function

```{r}

# df is data frame of specific subsector
# mod is the model that was run

createBaseline <- function( df, mod )
{
  
  # only IVs in the model

  td <- df[ , c("Accrual","GovtMoneyRat","HHI","UNAgrand","FixedCostRat",
                "SurplusRat_ndrop_w892","EqRat_w_K","ProfFundFeeYes")       ]
  
  
  period1 <- c( c(1,0,0,0,0,0), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  period2 <- c( c(0,1,0,0,0,0), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  period3 <- c( c(0,0,1,0,0,0), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  period4 <- c( c(0,0,0,1,0,0), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  period5 <- c( c(0,0,0,0,1,0), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  period6 <- c( c(0,0,0,0,0,1), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  
  # period1[12] <- median( dat$SurplusRat_ndrop_w892, na.rm=T )
  # period2[12] <- median( dat$SurplusRat_ndrop_w892, na.rm=T )
  # period3[12] <- median( dat$SurplusRat_ndrop_w892, na.rm=T )
  # period4[12] <- median( dat$SurplusRat_ndrop_w892, na.rm=T )
  # period5[12] <- median( dat$SurplusRat_ndrop_w892, na.rm=T )
  # period6[12] <- median( dat$SurplusRat_ndrop_w892, na.rm=T )
  
  # period1 <- c(1,0,0,0,0,0, 0,0,0.54,0,0,-0.49853,1.9,0)
  # period2 <- c(0,1,0,0,0,0, 0,0,0.54,0,0,-0.49853,1.9,0)
  # period3 <- c(0,0,1,0,0,0, 0,0,0.54,0,0,-0.49853,1.9,0)
  # period4 <- c(0,0,0,1,0,0, 0,0,0.54,0,0,-0.49853,1.9,0)
  # period5 <- c(0,0,0,0,1,0, 0,0,0.54,0,0,-0.49853,1.9,0)
  # period6 <- c(0,0,0,0,0,1, 0,0,0.54,0,0,-0.49853,1.9,0)
  
  
  h1 <- (    1 / ( 1 + exp(-( sum(period1 * mod$coefficients) ))   )     )
  h2 <- (    1 / ( 1 + exp(-( sum(period2 * mod$coefficients) ))   )     )
  h3 <- (    1 / ( 1 + exp(-( sum(period3 * mod$coefficients) ))   )     )
  h4 <- (    1 / ( 1 + exp(-( sum(period4 * mod$coefficients) ))   )     )
  h5 <- (    1 / ( 1 + exp(-( sum(period5 * mod$coefficients) ))   )     )
  h6 <- (    1 / ( 1 + exp(-( sum(period6 * mod$coefficients) ))   )     )
  
  hrc <- data.frame( age=1:6, hazard.rate=round(c(h1,h2,h3,h4,h5,h6),4) ) 
  
  src <- cumprod( (1-hrc$hazard.rate))
  
  hrc <- cbind( hrc, survival.rate=src )
  
  hrc
  
}


```


## Baseline Rates

```{r, results='asis'}

hrc <- createBaseline( df=dat.original, mod=m.all )

# dput( hrc )

# hrt <- calc.effect( model=m.all, df=dat, variable.name="Accrual", quant=1 ) 

# hrc %>% pander

saveRDS( hrc, "./Data/BaselineVitalRates.rds" )

stargazer( hrc, summary=FALSE, type="html", out="./Results/BaseVitalRates.doc" )

```


## Calc Effect Function

```{r}


# prof ~ D1 + D2 + D3 + D4 + D5 + D6 + Accrual + GovtMoneyRat + 
#     HHI + UNAgrand + FixedCostRat + SurplusRat_ndrop_w892 + EqRat_w_K + 
#     ProfFundFeeYes - 1


# CALC EFFECTS AT 25th and 75th percentiles, or 0 to 1 for binary
# Accrual - 0 to 1
# GovMoneyRat - 0 to 0.5
# HHI 0.54 to 0.99
# UNAgrand 0 to 44.8
# FixedCostRat 0 to 0.06221
# SurplusRat_ndrop_w892  -0.49853  to  0.14060
# EqRat_w_K  1.9  to  24912
# ProfFundFeeYes 0 to 1


calc.effect <- function( model, df, variable.name, quant )
{

  td <- df[ , c("Accrual","GovtMoneyRat","HHI","UNAgrand","FixedCostRat",
                "SurplusRat_ndrop_w892","EqRat_w_K","ProfFundFeeYes")         ]
  
  period1 <- c( c(1,0,0,0,0,0), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  period2 <- c( c(0,1,0,0,0,0), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  period3 <- c( c(0,0,1,0,0,0), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  period4 <- c( c(0,0,0,1,0,0), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  period5 <- c( c(0,0,0,0,1,0), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  period6 <- c( c(0,0,0,0,0,1), unlist( lapply( td, quantile, 0.50, na.rm=T ) ) )
  
  
  coef.names <- c("D1","D2","D3","D4","D5","D6","Accrual","GovtMoneyRat", 
                 "HHI","UNAgrand","FixedCostRat","SurplusRat_ndrop_w892", 
                 "EqRat_w_K","ProfFundFeeYes" )
  
  this.one <- which( coef.names == variable.name )
  
  new.value <- quantile( df[,variable.name], quant, na.rm=T )
  
  period1[ this.one ] <- new.value
  period2[ this.one ] <- new.value
  period3[ this.one ] <- new.value
  period4[ this.one ] <- new.value
  period5[ this.one ] <- new.value
  period6[ this.one ] <- new.value

  h1 <- (    1 / ( 1 + exp(-( sum(period1 * model$coefficients) ))   )     )
  h2 <- (    1 / ( 1 + exp(-( sum(period2 * model$coefficients) ))   )     )
  h3 <- (    1 / ( 1 + exp(-( sum(period3 * model$coefficients) ))   )     )
  h4 <- (    1 / ( 1 + exp(-( sum(period4 * model$coefficients) ))   )     )
  h5 <- (    1 / ( 1 + exp(-( sum(period5 * model$coefficients) ))   )     )
  h6 <- (    1 / ( 1 + exp(-( sum(period6 * model$coefficients) ))   )     )
  
  hrt <- data.frame( age=1:6, hazard.rate=round( c(h1,h2,h3,h4,h5,h6), 4) ) 
  
  srt <- cumprod( (1-hrt$hazard.rate))
  
  hrt <- cbind( hrt, survival.rate=srt )
  
  hrt
  
}



```



## Example Effect

```{r}


# "ACCRUAL"
hrt <- calc.effect( model=m.all, df=dat, variable.name="Accrual", quant=1 )

saveRDS( hrt, "./Data/AdjustedVitalRates.rds" )

# dput( hrt )
```




## Examine Effect Sizes of Each Variable

```{r}

# prof ~ D1 + D2 + D3 + D4 + D5 + D6 + Accrual + GovtMoneyRat + 
#     HHI + UNAgrand + FixedCostRat + SurplusRat_ndrop_w892 + EqRat_w_K + 
#     ProfFundFeeYes - 1


# CALC EFFECTS AT 25th and 75th percentiles, or 0 to 1 for binary
# Accrual - 0 to 1
# GovMoneyRat - 0 to 0.5
# HHI 0.54 to 0.99
# UNAgrand 0 to 44.8
# FixedCostRat 0 to 0.06221
# SurplusRat_ndrop_w892  -0.49853  to  0.14060
# EqRat_w_K  1.9  to  24912
# ProfFundFeeYes 0 to 1



# "BASELINE MODEL"
hrc %>% kable( caption="BASELINE MODEL" )


# "ACCRUAL"
quantile( dat$Accrual, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
calc.effect( model=m.all, df=dat, variable.name="Accrual", quant=1 ) %>% kable( caption="ACCRUAL" )


# "GovtMoneyRat"
quantile( dat$GovtMoneyRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
calc.effect( model=m.all, df=dat, variable.name="GovtMoneyRat", quant=0.99 ) %>% kable( caption="GovtMoneyRat" )


# "HHI"
quantile( dat$HHI, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
calc.effect( model=m.all, df=dat, variable.name="HHI", quant=0.99 ) %>% kable( caption="HHI" )


#  "UNAgrand" 
quantile( dat$UNAgrand, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
calc.effect( model=m.all, df=dat, variable.name="UNAgrand", quant=0.99 ) %>% kable( caption="UNAgrand" )


# "FixedCostRat"
quantile( dat$FixedCostRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
calc.effect( model=m.all, df=dat, variable.name="FixedCostRat", quant=0.99 ) %>% kable( caption="FixedCostRat" )


# "SurplusRat_ndrop_w892"
quantile( dat$SurplusRat_ndrop_w892, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
calc.effect( model=m.all, df=dat, variable.name="SurplusRat_ndrop_w892", quant=0.99 ) %>% kable( caption="SURPLUS RATIO" )


#"EqRat_w_K"
quantile( dat$EqRat_w_K, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
calc.effect( model=m.all, df=dat, variable.name="EqRat_w_K", quant=0.99 ) %>% kable( caption="EQUITY RATIO" )


# "ProfFundFeeYes"
quantile( dat$ProfFundFeeYes, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
calc.effect( model=m.all, df=dat, variable.name="ProfFundFeeYes", quant=1 ) %>% kable( caption="PROF FUNDRAISING?" )



```



## Plot of Effects

```{r}


# "BASELINE MODEL"
# hrc %>% kable( caption="BASELINE MODEL" )


# "ACCRUAL"
quantile( dat$Accrual, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.accrual <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="Accrual", quant=1 )[,"survival.rate"]


# "GovMoneyRat"
quantile( dat$GovtMoneyRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.gov <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="GovtMoneyRat", quant=0.99 )[,"survival.rate"]


# "HHI"
quantile( dat$HHI, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.hhi <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="HHI", quant=0.99 )[,"survival.rate"]


#  "UNAgrand" 
quantile( dat$UNAgrand, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.una <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="UNAgrand", quant=0.99 )[,"survival.rate"]


# "FixedCostRat"
quantile( dat$FixedCostRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fcr <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="FixedCostRat", quant=0.99 )[,"survival.rate"]


# "SurplusRat_ndrop_w892" 
quantile( dat$SurplusRat_ndrop_w892, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.surp <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="SurplusRat_ndrop_w892", quant=0.99 )[,"survival.rate"]


#"EqRat_w_K"
quantile( dat$EqRat_w_K, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.eqrat <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="EqRat_w_K", quant=0.99 )[,"survival.rate"]


# "ProfFundFeeYes"
quantile( dat$ProfFundFeeYes, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fundfee <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="ProfFundFeeYes", quant=1 )[,"survival.rate"]







plot( 1:6, rep(0,6), type="n", ylim=c(-0.1,0.5), bty="n",
      ylab="Increase in Prob. of Professionalization", xlab="Time in Years")
abline( h=0, lty=2, col="red" )

points( 1:6, effect.accrual, type="b", pch=19, cex=0.5 )
points( 1:6, effect.gov, type="b", pch=19, cex=0.5 )
points( 1:6, effect.hhi, type="b", pch=19, cex=0.5 )
points( 1:6, effect.una, type="b", pch=19, cex=0.5 )
points( 1:6, effect.fcr, type="b", pch=19, cex=0.5 )
points( 1:6, effect.surp, type="b", pch=19, cex=0.5 )
points( 1:6, effect.eqrat, type="b", pch=19, cex=0.5 )
points( 1:6, effect.fundfee, type="b", pch=19, cex=0.5 )



Effect.Size <- data.frame(
  prob.increase=c( effect.accrual[6], effect.gov[6], effect.hhi[6], effect.una[6], 
     effect.fcr[6], effect.surp[6], effect.eqrat[6], effect.fundfee[6] ),
  var.name=c("Accrual Accounting","Government Revenue","Revenue Concentration",
    "Unrestricted Net Assets","Fixed Cost Ratios","Surplus Ratio",
    "Equity Ratio","Professional Fundraising")
)

Effect.Size 

this.order <- order( Effect.Size$prob.increase )  # set one order for all graphs

ES.all <- Effect.Size[ this.order , ]



Effect.Size <- arrange( Effect.Size, prob.increase )



dotchart( Effect.Size$prob.increase, labels=Effect.Size$var.name, pch=19, bty="n" )

```



```{r, echo=F}

# p <- plot_ly( Effect.Size , x=prob.increase, y=var.name, mode = "markers", marker=list(color="gray", size=20) ) 
#   
# layout(
#          title = "Effect Size of Each Study Variable",
#          xaxis = list(title = "Increase in Probability of Professionalization"),
#          yaxis = list(title = ""),
#          margin = list(l = 200)
#   )
# 
# p
```





# ARTS

```{r}

dat <- dat.original[ dat.original$Subsector2 == "Arts" ,  ]

# > unique( as.character(dat$Subsector2) )
# [1] "Edu(Exclu Higher)"       "Human Svcs"             
# [3]                     "Other"                  
# [5] "Public"                  "Health"                 
# [7] "Higher Edu or Hospitals"


m.arts <- glm( formula= prof ~ D1 + D2 + D3 + D4 + D5 + D6 + Accrual + GovtMoneyRat + HHI + 
                             UNAgrand + FixedCostRat + SurplusRat_ndrop_w892 + EqRat_w_K + 
                             ProfFundFeeYes - 1, 
                             data=dat,   family=binomial("logit") )

m.arts %>% tidy %>% kable( digits=3 )



# "BASELINE MODEL"
hrc <- createBaseline( df=dat, mod=m.arts )
hrc %>% pander


# "ACCRUAL"
quantile( dat$Accrual, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.accrual <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="Accrual", quant=1 )[,"survival.rate"]


# "GovMoneyRat"
quantile( dat$GovtMoneyRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.gov <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="GovtMoneyRat", quant=0.99 )[,"survival.rate"]


# "HHI"
quantile( dat$HHI, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.hhi <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="HHI", quant=0.99 )[,"survival.rate"]


#  "UNAgrand" 
quantile( dat$UNAgrand, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.una <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="UNAgrand", quant=0.99 )[,"survival.rate"]


# "FixedCostRat"
quantile( dat$FixedCostRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fcr <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="FixedCostRat", quant=0.99 )[,"survival.rate"]


# "SurplusRat_ndrop_w892" 
quantile( dat$SurplusRat_ndrop_w892, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.surp <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="SurplusRat_ndrop_w892", quant=0.99 )[,"survival.rate"]


#"EqRat_w_K"
quantile( dat$EqRat_w_K, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.eqrat <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="EqRat_w_K", quant=0.99 )[,"survival.rate"]


# "ProfFundFeeYes"
quantile( dat$ProfFundFeeYes, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fundfee <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="ProfFundFeeYes", quant=1 )[,"survival.rate"]




Effect.Size <- data.frame(
  prob.increase=c( effect.accrual[6], effect.gov[6], effect.hhi[6], effect.una[6], 
     effect.fcr[6], effect.surp[6], effect.eqrat[6], effect.fundfee[6] ),
  var.name=c("Accrual Accounting","Government Revenue","Revenue Concentration",
    "Unrestricted Net Assets","Fixed Cost Ratios","Surplus Ratio",
    "Equity Ratio","Professional Fundraising")
)



Effect.Size 

ES.arts <- Effect.Size[ this.order , ]

Effect.Size <- arrange( Effect.Size, prob.increase )



#dotchart( Effect.Size$prob.increase, labels=Effect.Size$var.name, pch=19, bty="n" )


# p <- plot_ly( Effect.Size , x=prob.increase, y=var.name, mode = "markers", marker=list(color="gray", size=20) ) 
#   
# layout(
#          title = "Effect Size of Each Study Variable",
#          xaxis = list(title = "Increase in Probability of Professionalization"),
#          yaxis = list(title = ""),
#          margin = list(l = 200)
#   )

# p


```






# HEALTH

```{r}

dat <- dat.original[ dat.original$Subsector2 == "Health" ,  ]

# > unique( as.character(dat$Subsector2) )
# [1] "Edu(Exclu Higher)"       "Human Svcs"             
# [3]                     "Other"                  
# [5] "Public"                                   
# [7] "Higher Edu or Hospitals"


m.health <- glm( formula= prof ~ D1 + D2 + D3 + D4 + D5 + D6 + Accrual + GovtMoneyRat + HHI + 
                             UNAgrand + FixedCostRat + SurplusRat_ndrop_w892 + EqRat_w_K + 
                             ProfFundFeeYes - 1, 
                             data=dat,   family=binomial("logit") )

m.health %>% tidy %>% kable( digits=3 )



# "BASELINE MODEL"
hrc <- createBaseline( df=dat, mod=m.health )
hrc %>% pander


# "ACCRUAL"
quantile( dat$Accrual, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.accrual <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="Accrual", quant=1 )[,"survival.rate"]


# "GovMoneyRat"
quantile( dat$GovtMoneyRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.gov <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="GovtMoneyRat", quant=0.99 )[,"survival.rate"]


# "HHI"
quantile( dat$HHI, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.hhi <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="HHI", quant=0.99 )[,"survival.rate"]


#  "UNAgrand" 
quantile( dat$UNAgrand, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.una <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="UNAgrand", quant=0.99 )[,"survival.rate"]


# "FixedCostRat"
quantile( dat$FixedCostRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fcr <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="FixedCostRat", quant=0.99 )[,"survival.rate"]


# "SurplusRat_ndrop_w892" 
quantile( dat$SurplusRat_ndrop_w892, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.surp <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="SurplusRat_ndrop_w892", quant=0.99 )[,"survival.rate"]


#"EqRat_w_K"
quantile( dat$EqRat_w_K, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.eqrat <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="EqRat_w_K", quant=0.99 )[,"survival.rate"]


# "ProfFundFeeYes"
quantile( dat$ProfFundFeeYes, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fundfee <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="ProfFundFeeYes", quant=1 )[,"survival.rate"]





Effect.Size <- data.frame(
  prob.increase=c( effect.accrual[6], effect.gov[6], effect.hhi[6], effect.una[6], 
     effect.fcr[6], effect.surp[6], effect.eqrat[6], effect.fundfee[6] ),
  var.name=c("Accrual Accounting","Government Revenue","Revenue Concentration",
    "Unrestricted Net Assets","Fixed Cost Ratios","Surplus Ratio",
    "Equity Ratio","Professional Fundraising")
)

Effect.Size 

ES.health <- Effect.Size[ this.order , ]


Effect.Size <- arrange( Effect.Size, prob.increase )




#dotchart( Effect.Size$prob.increase, labels=Effect.Size$var.name, pch=19, bty="n" )


# p <- plot_ly( Effect.Size , x=prob.increase, y=var.name, mode = "markers", marker=list(color="gray", size=20) ) 
#   
# layout(
#          title = "Effect Size of Each Study Variable",
#          xaxis = list(title = "Increase in Probability of Professionalization"),
#          yaxis = list(title = ""),
#          margin = list(l = 200)
#   )

# p


```




# HUMAN SERVICES

```{r}

dat <- dat.original[ dat.original$Subsector2 == "Human Svcs" ,  ]

# > unique( as.character(dat$Subsector2) )
# [1] "Edu(Exclu Higher)"                    
# [3]                     "Other"                  
# [5] "Public"                                   
# [7] "Higher Edu or Hospitals"


m.hs <- glm( formula= prof ~ D1 + D2 + D3 + D4 + D5 + D6 + Accrual + GovtMoneyRat + HHI + 
                             UNAgrand + FixedCostRat + SurplusRat_ndrop_w892 + EqRat_w_K + 
                             ProfFundFeeYes - 1, 
                             data=dat,   family=binomial("logit") )

m.hs %>% tidy %>% kable( digits=3 )



# "BASELINE MODEL"
hrc <- createBaseline( df=dat, mod=m.hs )
hrc %>% pander


# "ACCRUAL"
quantile( dat$Accrual, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.accrual <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="Accrual", quant=1 )[,"survival.rate"]


# "GovMoneyRat"
quantile( dat$GovtMoneyRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.gov <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="GovtMoneyRat", quant=0.99 )[,"survival.rate"]


# "HHI"
quantile( dat$HHI, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.hhi <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="HHI", quant=0.99 )[,"survival.rate"]


#  "UNAgrand" 
quantile( dat$UNAgrand, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.una <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="UNAgrand", quant=0.99 )[,"survival.rate"]


# "FixedCostRat"
quantile( dat$FixedCostRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fcr <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="FixedCostRat", quant=0.99 )[,"survival.rate"]


# "SurplusRat_ndrop_w892" 
quantile( dat$SurplusRat_ndrop_w892, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.surp <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="SurplusRat_ndrop_w892", quant=0.99 )[,"survival.rate"]


#"EqRat_w_K"
quantile( dat$EqRat_w_K, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.eqrat <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="EqRat_w_K", quant=0.99 )[,"survival.rate"]


# "ProfFundFeeYes"
quantile( dat$ProfFundFeeYes, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fundfee <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="ProfFundFeeYes", quant=1 )[,"survival.rate"]






Effect.Size <- data.frame(
  prob.increase=c( effect.accrual[6], effect.gov[6], effect.hhi[6], effect.una[6], 
     effect.fcr[6], effect.surp[6], effect.eqrat[6], effect.fundfee[6] ),
  var.name=c("Accrual Accounting","Government Revenue","Revenue Concentration",
    "Unrestricted Net Assets","Fixed Cost Ratios","Surplus Ratio",
    "Equity Ratio","Professional Fundraising")
)

Effect.Size 

ES.hs <- Effect.Size[ this.order , ]


Effect.Size <- arrange( Effect.Size, prob.increase )







#dotchart( Effect.Size$prob.increase, labels=Effect.Size$var.name, pch=19, bty="n" )


p <- plot_ly( Effect.Size , x=~prob.increase, y=~var.name, type="scatter", 
              mode="marker", marker=list(color="gray", size=20) ) %>%

layout(
         title = "Effect Size of Each Study Variable",
         xaxis = list(title = "Increase in Probability of Professionalization"),
         yaxis = list(title = ""),
         margin = list(l = 200)
  )

p


```






# PUBLIC


```{r}

dat <- dat.original[ dat.original$Subsector2 == "Public" ,  ]

# > unique( as.character(dat$Subsector2) )
# [1] "Edu(Exclu Higher)"                    
# [3]                     "Other"                  
# [5]                                    
# [7] "Higher Edu or Hospitals"


m.pub <- glm( formula= prof ~ D1 + D2 + D3 + D4 + D5 + D6 + Accrual + GovtMoneyRat + HHI + 
                             UNAgrand + FixedCostRat + SurplusRat_ndrop_w892 + EqRat_w_K + 
                             ProfFundFeeYes - 1, 
                             data=dat,   family=binomial("logit") )

m.pub %>% tidy %>% kable( digits=3 )



# "BASELINE MODEL"
hrc <- createBaseline( df=dat, mod=m.pub )
hrc %>% pander


# "ACCRUAL"
quantile( dat$Accrual, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.accrual <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="Accrual", quant=1 )[,"survival.rate"]


# "GovMoneyRat"
quantile( dat$GovtMoneyRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.gov <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="GovtMoneyRat", quant=0.99 )[,"survival.rate"]


# "HHI"
quantile( dat$HHI, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.hhi <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="HHI", quant=0.99 )[,"survival.rate"]


#  "UNAgrand" 
quantile( dat$UNAgrand, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.una <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="UNAgrand", quant=0.99 )[,"survival.rate"]


# "FixedCostRat"
quantile( dat$FixedCostRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fcr <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="FixedCostRat", quant=0.99 )[,"survival.rate"]


# "SurplusRat_ndrop_w892" 
quantile( dat$SurplusRat_ndrop_w892, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.surp <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="SurplusRat_ndrop_w892", quant=0.99 )[,"survival.rate"]


#"EqRat_w_K"
quantile( dat$EqRat_w_K, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.eqrat <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="EqRat_w_K", quant=0.99 )[,"survival.rate"]


# "ProfFundFeeYes"
quantile( dat$ProfFundFeeYes, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fundfee <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="ProfFundFeeYes", quant=1 )[,"survival.rate"]






Effect.Size <- data.frame(
  prob.increase=c( effect.accrual[6], effect.gov[6], effect.hhi[6], effect.una[6], 
     effect.fcr[6], effect.surp[6], effect.eqrat[6], effect.fundfee[6] ),
  var.name=c("Accrual Accounting","Government Revenue","Revenue Concentration",
    "Unrestricted Net Assets","Fixed Cost Ratios","Surplus Ratio",
    "Equity Ratio","Professional Fundraising")
)

Effect.Size 

ES.pub <- Effect.Size[ this.order , ]

Effect.Size <- arrange( Effect.Size, prob.increase )






#dotchart( Effect.Size$prob.increase, labels=Effect.Size$var.name, pch=19, bty="n" )


# p <- plot_ly( Effect.Size , x=prob.increase, y=var.name, mode = "markers", marker=list(color="gray", size=20) ) 
#   
# layout(
#          title = "Effect Size of Each Study Variable",
#          xaxis = list(title = "Increase in Probability of Professionalization"),
#          yaxis = list(title = ""),
#          margin = list(l = 200)
#   )

# p


```




# EDUCATION


```{r}

dat <- dat.original[ dat.original$Subsector2 == "Edu(Exclu Higher)" ,  ]

# > unique( as.character(dat$Subsector2) )
# [1]                     
# [3]                     "Other"                  
# [5]                                    
# [7] "Higher Edu or Hospitals"


m.edu <- glm( formula= prof ~ D1 + D2 + D3 + D4 + D5 + D6 + Accrual + GovtMoneyRat + HHI + 
                             UNAgrand + FixedCostRat + SurplusRat_ndrop_w892 + EqRat_w_K + 
                             ProfFundFeeYes - 1, 
                             data=dat,   family=binomial("logit") )

m.edu %>% tidy %>% kable( digits=3 )



# "BASELINE MODEL"
hrc <- createBaseline( df=dat, mod=m.edu )
hrc %>% pander


# "ACCRUAL"
quantile( dat$Accrual, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.accrual <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="Accrual", quant=1 )[,"survival.rate"]


# "GovMoneyRat"
quantile( dat$GovtMoneyRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.gov <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="GovtMoneyRat", quant=0.99 )[,"survival.rate"]


# "HHI"
quantile( dat$HHI, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.hhi <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="HHI", quant=0.99 )[,"survival.rate"]


#  "UNAgrand" 
quantile( dat$UNAgrand, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.una <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="UNAgrand", quant=0.99 )[,"survival.rate"]


# "FixedCostRat"
quantile( dat$FixedCostRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fcr <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="FixedCostRat", quant=0.99 )[,"survival.rate"]


# "SurplusRat_ndrop_w892" 
quantile( dat$SurplusRat_ndrop_w892, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.surp <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="SurplusRat_ndrop_w892", quant=0.99 )[,"survival.rate"]


#"EqRat_w_K"
quantile( dat$EqRat_w_K, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.eqrat <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="EqRat_w_K", quant=0.99 )[,"survival.rate"]


# "ProfFundFeeYes"
quantile( dat$ProfFundFeeYes, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fundfee <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="ProfFundFeeYes", quant=1 )[,"survival.rate"]






Effect.Size <- data.frame(
  prob.increase=c( effect.accrual[6], effect.gov[6], effect.hhi[6], effect.una[6], 
     effect.fcr[6], effect.surp[6], effect.eqrat[6], effect.fundfee[6] ),
  var.name=c("Accrual Accounting","Government Revenue","Revenue Concentration",
    "Unrestricted Net Assets","Fixed Cost Ratios","Surplus Ratio",
    "Equity Ratio","Professional Fundraising")
)

Effect.Size 

ES.edu <- Effect.Size[ this.order , ]


Effect.Size <- arrange( Effect.Size, prob.increase )





#dotchart( Effect.Size$prob.increase, labels=Effect.Size$var.name, pch=19, bty="n" )


# p <- plot_ly( Effect.Size , x=prob.increase, y=var.name, mode = "markers", marker=list(color="gray", size=20) ) 
#   
# layout(
#          title = "Effect Size of Each Study Variable",
#          xaxis = list(title = "Increase in Probability of Professionalization"),
#          yaxis = list(title = ""),
#          margin = list(l = 200)
#   )

# p


```




# OTHER


```{r}

dat <- dat.original[ dat.original$Subsector2 == "Other" ,  ]

# > unique( as.character(dat$Subsector2) )
# [1] "Edu(Exclu Higher)"                    
# [3]                                       
# [5]                                    
# [7] "Higher Edu or Hospitals"


m.other <- glm( formula= prof ~ D1 + D2 + D3 + D4 + D5 + D6 + Accrual + GovtMoneyRat + HHI + 
                             UNAgrand + FixedCostRat + SurplusRat_ndrop_w892 + EqRat_w_K + 
                             ProfFundFeeYes - 1, 
                             data=dat,   family=binomial("logit") )

m.other %>% tidy %>% kable( digits=3 )



# "BASELINE MODEL"
hrc <- createBaseline( df=dat, mod=m.other )
hrc %>% pander


# "ACCRUAL"
quantile( dat$Accrual, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.accrual <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="Accrual", quant=1 )[,"survival.rate"]


# "GovMoneyRat"
quantile( dat$GovtMoneyRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.gov <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="GovtMoneyRat", quant=0.99 )[,"survival.rate"]


# "HHI"
quantile( dat$HHI, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.hhi <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="HHI", quant=0.99 )[,"survival.rate"]


#  "UNAgrand" 
quantile( dat$UNAgrand, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.una <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="UNAgrand", quant=0.99 )[,"survival.rate"]


# "FixedCostRat"
quantile( dat$FixedCostRat, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fcr <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="FixedCostRat", quant=0.99 )[,"survival.rate"]


# "SurplusRat_ndrop_w892" 
quantile( dat$SurplusRat_ndrop_w892, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.surp <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="SurplusRat_ndrop_w892", quant=0.99 )[,"survival.rate"]


#"EqRat_w_K"
quantile( dat$EqRat_w_K, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.eqrat <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="EqRat_w_K", quant=0.99 )[,"survival.rate"]


# "ProfFundFeeYes"
quantile( dat$ProfFundFeeYes, c(0.05,0.25,0.5,0.75,0.95,0.99,1) )
effect.fundfee <- hrc$survival.rate - calc.effect( model=m.all, df=dat, variable.name="ProfFundFeeYes", quant=1 )[,"survival.rate"]






Effect.Size <- data.frame(
  prob.increase=c( effect.accrual[6], effect.gov[6], effect.hhi[6], effect.una[6], 
     effect.fcr[6], effect.surp[6], effect.eqrat[6], effect.fundfee[6] ),
  var.name=c("Accrual Accounting","Government Revenue","Revenue Concentration",
    "Unrestricted Net Assets","Fixed Cost Ratios","Surplus Ratio",
    "Equity Ratio","Professional Fundraising")
)

Effect.Size 

ES.other <- Effect.Size[ this.order , ]


Effect.Size <- arrange( Effect.Size, prob.increase )






#dotchart( Effect.Size$prob.increase, labels=Effect.Size$var.name, pch=19, bty="n" )


# p <- plot_ly( Effect.Size , x=prob.increase, y=var.name, mode = "markers", marker=list(color="gray", size=20) ) 
#   
# layout(
#          title = "Effect Size of Each Study Variable",
#          xaxis = list(title = "Increase in Probability of Professionalization"),
#          yaxis = list(title = ""),
#          margin = list(l = 200)
#   )

# p


```









# ALL MODELS



```{r, results='asis'}


# mtable123 <- mtable( "Combined"=m.all, "Arts"=m.arts, "Health"=m.health,
#                      "Human Services"=m.hs, "Public"=m.edu, "Education"=m.other, 
#                      "Misc."=m.pub,
#                      summary.stats=c("p","N") )
# 
# 
# 
# panderOptions("digits", 2)
# mtable123 # %>% pander


stargazer( m.all, m.arts, m.health, m.hs, m.pub, m.edu, m.other, type="html", digits=2,
           column.labels=c("All","Arts","Health","Human Services","Public","Education","Misc"),
           title="Discrete Hazard Model Estimates",
           dep.var.labels=c("Professionalization"),
           no.space = TRUE
         )


```




# TABLE OF EFFECTS



```{r}


# sanity check to make sure rows align correctly

ef.tab <-   cbind(        ALL=ES.all, 
                          ARTS=ES.arts, 
                          HEALTH=ES.health, 
                          HUMAN_SERVICES=ES.hs, 
                          PUBLIC=ES.pub, 
                          EDUCATION=ES.edu, 
                          OTHER=ES.other)

as.data.frame(ef.tab) %>% pander




ef.table <- round( cbind( ALL=ES.all$prob.increase, 
                          ARTS=ES.arts$prob.increase, 
                          HEALTH=ES.health$prob.increase, 
                          HUMAN_SERVICES=ES.hs$prob.increase, 
                          PUBLIC=ES.pub$prob.increase, 
                          EDUCATION=ES.edu$prob.increase, 
                          OTHER=ES.other$prob.increase), 3  ) 

ef.table <- as.data.frame( ef.table )

row.names(ef.table) <- as.character(ES.all$var.name)

ef.table <- ef.table[ 8:1 , ]

# ef.table %>% pander

saveRDS( ef.table, "./Data/EffectsTable.rds" )


```


## Stargazer of Effects Table

```{r, results='asis'}

stargazer( ef.table, summary=FALSE, type="html", out="./Results/EffectsTable2.doc", digits=2 )

```
